# ascii/valid for arm64