--- a/ascii/valid_arm64.go
+++ b/ascii/valid_arm64.go
@@ -1,0 +1,7 @@
+//go:build !purego && arm64
+// +build !purego,arm64
+
+package ascii
+
+// ValidString returns true if s contains only ASCII characters.
+func ValidString(s string) bool