--- a/ascii/valid_arm64.go
+++ b/ascii/valid_arm64.go
@@ -1,5 +1,5 @@
-//go:build !purego && arm64
-// +build !purego,arm64
+//go:build !purego
+// +build !purego
 
 package ascii