… ascii/valid_amd64.go