Forgot to open this PR following the hack week. Benchmark reveal that this implementation is slower with small strings. I will have to disassemble the native implementation to see what's happening in there.

We have good results with long strings since we are leveraging ASIMD there. However, during my tests I found that loading and testing against more than 32B at once was not granting meaningful performance gains. I'm sure I'm doing something wrong there.

Since we have access to SVE in Graviton 3, we could implement a more efficient looping system using the new `whilelt` instruction.

```
name                              old time/op    new time/op      delta
Valid/#00-10                        2.19ns ± 1%      2.82ns ± 1%     +28.74%  (p=0.008 n=5+5)
Valid/a-10                          2.23ns ± 2%      2.83ns ± 1%     +27.11%  (p=0.008 n=5+5)
Valid/ab-10                         2.56ns ± 1%      2.53ns ± 1%        ~     (p=0.111 n=5+5)
Valid/abc-10                        2.50ns ± 1%      2.22ns ± 1%     -11.24%  (p=0.008 n=5+5)
Valid/abcd-10                       2.18ns ± 0%      3.83ns ± 1%     +75.60%  (p=0.008 n=5+5)
Valid/hello-10                      2.18ns ± 0%      3.78ns ± 1%     +73.20%  (p=0.008 n=5+5)
Valid/Hello_World!-10               2.49ns ± 0%      4.70ns ± 0%     +88.54%  (p=0.008 n=5+5)
Valid/Hello"World!-10               2.49ns ± 0%      4.68ns ± 0%     +87.96%  (p=0.008 n=5+5)
Valid/Hello\World!-10               2.49ns ± 0%      4.71ns ± 1%     +89.05%  (p=0.008 n=5+5)
Valid/Hello_World!#01-10            2.49ns ± 0%      4.75ns ± 1%     +90.66%  (p=0.008 n=5+5)
Valid/Hello_World!#02-10            2.49ns ± 0%      4.72ns ± 0%     +89.22%  (p=0.008 n=5+5)
Valid/Hello_World!#03-10            2.50ns ± 0%      4.71ns ± 0%     +88.84%  (p=0.008 n=5+5)
Valid/Hello\bWorld!-10              2.51ns ± 1%      4.73ns ± 0%     +88.59%  (p=0.008 n=5+5)
Valid/Hello_World!#04-10            2.50ns ± 0%      4.70ns ± 0%     +88.07%  (p=0.008 n=5+5)
Valid/H~llo_World!-10               2.50ns ± 0%      4.73ns ± 0%     +89.14%  (p=0.008 n=5+5)
Valid/H~llo-10                      2.18ns ± 0%      3.76ns ± 0%     +72.28%  (p=0.016 n=4+5)
Valid/你好-10                         2.05ns ± 1%      2.03ns ± 1%      -0.87%  (p=0.016 n=5+5)
Valid/~-10                          2.18ns ± 0%      2.84ns ± 0%     +30.42%  (p=0.008 n=5+5)
Valid/�-10                          2.18ns ± 0%      2.83ns ± 1%     +29.65%  (p=0.008 n=5+5)
Valid/\u007f-10                     2.18ns ± 0%      2.83ns ± 1%     +29.70%  (p=0.008 n=5+5)
Valid/�#01-10                       2.19ns ± 0%      2.84ns ± 0%     +29.64%  (p=0.008 n=5+5)
Valid/a_string_of_16B.-10           3.43ns ± 0%      2.25ns ± 5%     -34.37%  (p=0.008 n=5+5)
Valid/an_invalid_string...-10       4.05ns ± 0%      3.20ns ± 3%     -20.97%  (p=0.008 n=5+5)
Valid/some_kind_of_long...-10       4.68ns ± 0%      2.43ns ± 2%     -47.96%  (p=0.008 n=5+5)
Valid/some_kind_of_long...#01-10    4.68ns ± 1%      2.33ns ± 0%     -50.27%  (p=0.008 n=5+5)
Valid/12345678901234567...-10        404ns ± 0%         2ns ± 1%     -99.41%  (p=0.008 n=5+5)
ValidString/0007-10                 2.53ns ± 1%      3.12ns ± 0%     +23.37%  (p=0.008 n=5+5)
ValidString/0008-10                 2.54ns ± 1%      3.75ns ± 1%     +47.73%  (p=0.008 n=5+5)
ValidString/0009-10                 2.54ns ± 2%      3.76ns ± 0%     +47.94%  (p=0.008 n=5+5)
ValidString/0015-10                 2.84ns ± 2%      4.08ns ± 0%     +43.79%  (p=0.008 n=5+5)
ValidString/0016-10                 3.45ns ± 0%      2.19ns ± 0%     -36.45%  (p=0.016 n=4+5)
ValidString/0017-10                 3.43ns ± 2%      2.21ns ± 1%     -35.78%  (p=0.008 n=5+5)
ValidString/0031-10                 4.09ns ± 1%      2.20ns ± 1%     -46.21%  (p=0.008 n=5+5)
ValidString/0032-10                 4.10ns ± 0%      3.14ns ± 0%     -23.50%  (p=0.008 n=5+5)
ValidString/0033-10                 4.10ns ± 0%      3.14ns ± 0%     -23.51%  (p=0.008 n=5+5)
ValidString/0512-10                 22.9ns ± 0%       2.3ns ± 1%     -89.88%  (p=0.008 n=5+5)
ValidString/2000-10                 90.5ns ± 1%       2.3ns ± 1%     -97.43%  (p=0.008 n=5+5)

name                              old speed      new speed        delta
Valid/a-10                         449MB/s ± 2%     353MB/s ± 1%     -21.36%  (p=0.008 n=5+5)
Valid/ab-10                        782MB/s ± 1%     789MB/s ± 1%        ~     (p=0.095 n=5+5)
Valid/abc-10                      1.20GB/s ± 1%    1.35GB/s ± 1%     +12.66%  (p=0.008 n=5+5)
Valid/abcd-10                     1.83GB/s ± 0%    1.04GB/s ± 1%     -43.05%  (p=0.008 n=5+5)
Valid/hello-10                    2.29GB/s ± 0%    1.32GB/s ± 1%     -42.26%  (p=0.008 n=5+5)
Valid/Hello_World!-10             4.82GB/s ± 0%    2.55GB/s ± 0%     -46.96%  (p=0.008 n=5+5)
Valid/Hello"World!-10             4.82GB/s ± 0%    2.56GB/s ± 0%     -46.80%  (p=0.008 n=5+5)
Valid/Hello\World!-10             4.81GB/s ± 0%    2.55GB/s ± 1%     -47.10%  (p=0.008 n=5+5)
Valid/Hello_World!#01-10          4.82GB/s ± 0%    2.53GB/s ± 1%     -47.55%  (p=0.008 n=5+5)
Valid/Hello_World!#02-10          4.81GB/s ± 0%    2.54GB/s ± 0%     -47.15%  (p=0.008 n=5+5)
Valid/Hello_World!#03-10          4.81GB/s ± 0%    2.55GB/s ± 0%     -47.04%  (p=0.008 n=5+5)
Valid/Hello\bWorld!-10            4.79GB/s ± 1%    2.54GB/s ± 0%     -46.98%  (p=0.008 n=5+5)
Valid/Hello_World!#04-10          4.80GB/s ± 0%    2.55GB/s ± 0%     -46.83%  (p=0.008 n=5+5)
Valid/H~llo_World!-10             4.80GB/s ± 0%    2.54GB/s ± 0%     -47.13%  (p=0.008 n=5+5)
Valid/H~llo-10                    2.29GB/s ± 1%    1.33GB/s ± 0%     -41.81%  (p=0.008 n=5+5)
Valid/你好-10                       2.92GB/s ± 1%    2.95GB/s ± 1%      +0.87%  (p=0.016 n=5+5)
Valid/~-10                         459MB/s ± 0%     352MB/s ± 0%     -23.33%  (p=0.008 n=5+5)
Valid/�-10                         459MB/s ± 0%     354MB/s ± 1%     -22.86%  (p=0.008 n=5+5)
Valid/\u007f-10                    459MB/s ± 0%     354MB/s ± 1%     -22.91%  (p=0.008 n=5+5)
Valid/�#01-10                      457MB/s ± 0%     352MB/s ± 0%     -22.87%  (p=0.008 n=5+5)
Valid/a_string_of_16B.-10         4.67GB/s ± 0%    7.12GB/s ± 5%     +52.49%  (p=0.008 n=5+5)
Valid/an_invalid_string...-10     8.15GB/s ± 0%   10.32GB/s ± 3%     +26.56%  (p=0.008 n=5+5)
Valid/some_kind_of_long...-10     11.1GB/s ± 0%    21.4GB/s ± 2%     +92.21%  (p=0.008 n=5+5)
Valid/some_kind_of_long...#01-10  13.7GB/s ± 1%    27.5GB/s ± 0%    +101.08%  (p=0.008 n=5+5)
Valid/12345678901234567...-10     24.7GB/s ± 0%  4174.4GB/s ± 1%  +16785.33%  (p=0.008 n=5+5)
ValidString/0007-10               2.77GB/s ± 1%    2.25GB/s ± 0%     -18.94%  (p=0.008 n=5+5)
ValidString/0008-10               3.15GB/s ± 1%    2.13GB/s ± 1%     -32.31%  (p=0.008 n=5+5)
ValidString/0009-10               3.54GB/s ± 2%    2.39GB/s ± 0%     -32.41%  (p=0.008 n=5+5)
ValidString/0015-10               5.29GB/s ± 2%    3.68GB/s ± 0%     -30.46%  (p=0.008 n=5+5)
ValidString/0016-10               4.64GB/s ± 1%    7.29GB/s ± 0%     +56.98%  (p=0.008 n=5+5)
ValidString/0017-10               4.95GB/s ± 2%    7.71GB/s ± 1%     +55.71%  (p=0.008 n=5+5)
ValidString/0031-10               7.58GB/s ± 1%   14.09GB/s ± 1%     +85.92%  (p=0.008 n=5+5)
ValidString/0032-10               7.80GB/s ± 0%   10.19GB/s ± 0%     +30.71%  (p=0.008 n=5+5)
ValidString/0033-10               8.04GB/s ± 0%   10.51GB/s ± 0%     +30.74%  (p=0.008 n=5+5)
ValidString/0512-10               22.4GB/s ± 0%   221.2GB/s ± 1%    +888.15%  (p=0.008 n=5+5)
ValidString/2000-10               22.1GB/s ± 1%   860.9GB/s ± 1%   +3793.38%  (p=0.008 n=5+5)
```

- [ ] Find a fix for small strings
- [ ] Use SVE/SVE2 instead of ASIMD
- [ ] Add fuzzing